Refactor apiTokenSection
