+++ jsapp/js/account/security/password/passwordSection.module.scss
@@ -4,7 +4,7 @@
 
 .root {
   margin-top: sizes.$x20;
-  margin-bottom: sizes.$x20;
+  margin-bottom: sizes.$x60;
   padding-top: sizes.$x14;
   display: flex;
   border-top: sizes.$x1 solid;
